@@ -21,11 +21,15 @@
         "email": "php@graphaware.com"
     }],
     "require": {
-        "php": "^5.6 || ^7.0",
+        "php": "^7.0||^8.0",
         "doctrine/annotations": "^1.2",
         "doctrine/collections": "^1.3",
         "doctrine/common": "^2.6",
-        "graphaware/neo4j-php-client": "^4.6"
+        "laudis/neo4j-php-client": "^2.3",
+        "guzzlehttp/guzzle": "^7.4",
+        "nyholm/psr7": "^1.4",
+        "nyholm/psr7-server": "^1.0",
+        "kriswallsmith/buzz": "^1.2"
     },
     "autoload": {
         "psr-4": {
@@ -40,10 +44,9 @@
         }
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^1.0",
-        "phpunit/phpunit": "^5.2",
         "symfony/finder": "^3.0",
-        "symfony/stopwatch": "^3.0"
+        "symfony/stopwatch": "^3.0",
+        "phpunit/phpunit": "^9.5"
     },
     "extra": {
         "branch-alias": {
