@@ -11,7 +11,7 @@
 
 namespace GraphAware\Neo4j\OGM\Persister;
 
-use GraphAware\Common\Cypher\Statement;
+use Laudis\Neo4j\Databags\Statement;
 use GraphAware\Neo4j\OGM\Converters\Converter;
 use GraphAware\Neo4j\OGM\EntityManager;
 use GraphAware\Neo4j\OGM\Metadata\NodeEntityMetadata;
@@ -70,7 +70,7 @@ class EntityPersister
             }
         }
 
-        $query = sprintf('CREATE (n:%s) SET n += {properties}', $this->classMetadata->getLabel());
+        $query = sprintf('CREATE (n:%s) SET n += $properties', $this->classMetadata->getLabel());
         if (!empty($extraLabels)) {
             foreach ($extraLabels as $label) {
                 $query .= ' SET n:'.$label;
@@ -118,7 +118,7 @@ class EntityPersister
         }
         $id = $this->classMetadata->getIdValue($object);
 
-        $query = 'MATCH (n) WHERE id(n) = {id} SET n += {props}';
+        $query = 'MATCH (n) WHERE id(n) = $id SET n += $props';
         if (!empty($extraLabels)) {
             foreach ($extraLabels as $label) {
                 $query .= ' SET n:'.$label;
@@ -142,18 +142,18 @@ class EntityPersister
     public function refresh($id, $entity)
     {
         $label = $this->classMetadata->getLabel();
-        $query = sprintf('MATCH (n:%s) WHERE id(n) = {%s} RETURN n', $label, 'id');
+        $query = sprintf('MATCH (n:%s) WHERE id(n) = $%s RETURN n', $label, 'id');
         $result = $this->entityManager->getDatabaseDriver()->run($query, ['id' => $id]);
 
-        if ($result->size() > 0) {
-            $node = $result->firstRecord()->nodeValue('n');
+        if ($result->count() > 0) {
+            $node = $result->first()->get('n');
             $this->entityManager->getEntityHydrator($this->className)->refresh($node, $entity);
         }
     }
 
     public function getDetachDeleteQuery($object)
     {
-        $query = 'MATCH (n) WHERE id(n) = {id} DETACH DELETE n';
+        $query = 'MATCH (n) WHERE id(n) = $id DETACH DELETE n';
         $id = $this->classMetadata->getIdValue($object);
 
         return Statement::create($query, ['id' => $id]);
@@ -161,7 +161,7 @@ class EntityPersister
 
     public function getDeleteQuery($object)
     {
-        $query = 'MATCH (n) WHERE id(n) = {id} DELETE n';
+        $query = 'MATCH (n) WHERE id(n) = $id DELETE n';
         $id = $this->classMetadata->getIdValue($object);
 
         return Statement::create($query, ['id' => $id]);
