@@ -11,7 +11,7 @@
 
 namespace GraphAware\Neo4j\OGM\Persister;
 
-use GraphAware\Common\Cypher\Statement;
+use Laudis\Neo4j\Databags\Statement;
 use GraphAware\Neo4j\OGM\Metadata\RelationshipMetadata;
 
 class RelationshipPersister
@@ -38,11 +38,10 @@ class RelationshipPersister
 
         $relStringPart = sprintf($relString, $relationship->getType());
 
-        $query = 'MATCH (a), (b) WHERE id(a) = {ida} AND id(b) = {idb}
-        MERGE (a)'.$relStringPart.'(b)
-        RETURN id(r)';
+        $query = 'MATCH (a), (b) WHERE id(a) = $ida AND id(b) = $idb MERGE (a)'.$relStringPart.'(b) RETURN id(r)';
 
-        return Statement::create($query, ['ida' => $entityIdA, 'idb' => $entityIdB], 'rel_create_'.$relationship->getPropertyName().$entityIdA.'_'.$entityIdB);
+        return Statement::create($query, ['ida' => $entityIdA, 'idb' => $entityIdB]);
+        // return Statement::create($query, ['ida' => $entityIdA, 'idb' => $entityIdB], 'rel_create_'.$relationship->getPropertyName().$entityIdA.'_'.$entityIdB);
     }
 
     public function getDeleteRelationshipQuery($entityIdA, $entityIdB, RelationshipMetadata $relationship)
@@ -63,7 +62,7 @@ class RelationshipPersister
 
         $relStringPart = sprintf($relString, $relationship->getType());
 
-        $query = 'MATCH (a), (b) WHERE id(a) = {ida} AND id(b) = {idb}
+        $query = 'MATCH (a), (b) WHERE id(a) = $ida AND id(b) = $idb
         MATCH (a)'.$relStringPart.'(b)
         DELETE r';
 
