@@ -11,7 +11,7 @@
 
 namespace GraphAware\Neo4j\OGM\Proxy;
 
-use GraphAware\Common\Type\Node;
+use Laudis\Neo4j\Types\Node;
 use GraphAware\Neo4j\OGM\Metadata\RelationshipMetadata;
 
 class NodeCollectionInitializer extends SingleNodeInitializer
