@@ -20,7 +20,7 @@ use GraphAware\Neo4j\OGM\Tests\Integration\IntegrationTestCase;
  */
 class Issue21Test extends IntegrationTestCase
 {
-    public function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->clearDb();
