@@ -20,7 +20,7 @@ use GraphAware\Neo4j\OGM\Tests\Integration\Models\BooleanLabel\BlogPost;
  */
 class BooleanLabelTest extends IntegrationTestCase
 {
-    public function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->clearDb();
@@ -83,7 +83,7 @@ class BooleanLabelTest extends IntegrationTestCase
         $this->em->flush();
         $this->assertGraphExist('(b:BlogPost:Published {title:"Learn X"})');
         $result = $this->client->run('MATCH (n:BlogPost) RETURN count(n) AS c');
-        $this->assertSame(1, $result->firstRecord()->get('c'));
+        $this->assertSame(1, $result->first()->get('c'));
     }
 
     public function testLabelCanBeAddedAfterLoadAndCommit()
@@ -102,7 +102,7 @@ class BooleanLabelTest extends IntegrationTestCase
         $this->assertGraphExist('(b:BlogPost:Published {title:"Learn X"})');
         $this->em->flush();
         $result = $this->client->run('MATCH (n:BlogPost) RETURN count(n) AS c');
-        $this->assertSame(1, $result->firstRecord()->get('c'));
+        $this->assertSame(1, $result->first()->get('c'));
     }
 
     public function testLabelCanBeRemovedAfterCreateAndCommit()
@@ -120,7 +120,7 @@ class BooleanLabelTest extends IntegrationTestCase
         $this->em->flush();
         $this->assertGraphNotExist('(b:BlogPost:Published {title:"Learn X"})');
         $result = $this->client->run('MATCH (n:BlogPost) RETURN count(n) AS c');
-        $this->assertSame(1, $result->firstRecord()->get('c'));
+        $this->assertSame(1, $result->first()->get('c'));
     }
 
     public function testLabelCanBeRemovedAfterLoadAndCommit()
@@ -141,6 +141,6 @@ class BooleanLabelTest extends IntegrationTestCase
         $this->assertGraphExist('(b:BlogPost {title:"Learn X"})');
         $this->em->flush();
         $result = $this->client->run('MATCH (n:BlogPost) RETURN count(n) AS c');
-        $this->assertSame(1, $result->firstRecord()->get('c'));
+        $this->assertSame(1, $result->first()->get('c'));
     }
 }
