@@ -5,13 +5,15 @@ namespace GraphAware\Neo4j\OGM\Tests\Integration;
 use GraphAware\Neo4j\OGM\Query;
 use GraphAware\Neo4j\OGM\Tests\Integration\Models\MoviesDemo\Movie;
 use GraphAware\Neo4j\OGM\Tests\Integration\Models\MoviesDemo\Person;
+use Laudis\Neo4j\Types\CypherList;
+use Laudis\Neo4j\Types\CypherMap;
 
 /**
  * @group complex-query
  */
 class ComplexQueryResultTest extends IntegrationTestCase
 {
-    public function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->clearDb();
@@ -20,8 +22,7 @@ class ComplexQueryResultTest extends IntegrationTestCase
 
     public function testQueryReturningMap()
     {
-        $q = $this->em->createQuery('MATCH (n:Person)-[r:ACTED_IN]->(m)
-        RETURN n, {roles: r.roles, movie: m} AS actInfo LIMIT 2');
+        $q = $this->em->createQuery('MATCH (n:Person)-[r:ACTED_IN]->(m) RETURN n, {movie: m, roles: r.roles} AS actInfo LIMIT 2');
 
         $q->addEntityMapping('n', Person::class);
         $q->addEntityMapping('actInfo', null, Query::HYDRATE_MAP);
@@ -31,10 +32,15 @@ class ComplexQueryResultTest extends IntegrationTestCase
         $this->assertCount(2, $result);
         $row = $result[0];
 
+
+        /** @var CypherMap $actorInfo */
+        $actorInfo = $row['actInfo'];
         $this->assertInstanceOf(Person::class, $row['n']);
-        $this->assertInternalType('array', $row['actInfo']);
-        $this->assertInternalType('array', $row['actInfo']['roles']);
-        $this->assertInstanceOf(Movie::class, $row['actInfo']['movie']);
+        $this->assertInstanceOf(CypherMap::class, $actorInfo);
+        $this->assertTrue($actorInfo->hasKey('movie'));
+        $this->assertTrue($actorInfo->hasKey('roles'));
+        $this->assertInstanceOf(Movie::class, $actorInfo->get('movie'));
+        $this->assertInstanceOf(CypherList::class, $actorInfo->get('roles'));
     }
 
     public function testQueryReturningMapCollectionMixed()
@@ -51,7 +57,7 @@ class ComplexQueryResultTest extends IntegrationTestCase
 
         $result = $q->getResult();
         $this->assertCount(1, $result);
-        $this->assertInternalType('array', $result[0]['actorInfos']);
+        $this->assertIsArray($result[0]['actorInfos']);
         $this->assertInstanceOf(Movie::class, $result[0]['actorInfos'][0]['movie']);
         $this->assertCount(12, $result[0]['actorInfos']);
     }
@@ -69,10 +75,14 @@ class ComplexQueryResultTest extends IntegrationTestCase
 
         $this->assertCount(10, $result);
         $row = $result[0];
+        /** @var CypherMap $actorInfo */
+        $infos = $row['infos'];
         $this->assertInstanceOf(Person::class, $row['n']);
-        $this->assertInternalType('array', $row['infos']);
-        $this->assertEquals($row['infos']['score'], count($row['infos']['movies']));
-        $this->assertInstanceOf(Movie::class, $row['infos']['movies'][0]);
+        $this->assertInstanceOf(CypherMap::class, $infos);
+        $this->assertEquals($infos->get('score'), count($infos->get('movies')));
+        foreach ($infos->get('movies') as $movie) {
+            $this->assertInstanceOf(Movie::class, $movie);
+        }
     }
 
     public function testQueryReturningMapAsOnlyColumn()
@@ -88,10 +98,13 @@ class ComplexQueryResultTest extends IntegrationTestCase
 
         $this->assertCount(10, $result);
         $row = $result[0];
-        $this->assertInternalType('array', $row['infos']);
+        /** @var CypherMap $actorInfo */
+        $infos = $row['infos'];
         $this->assertEquals(1, count(array_keys($row)));
-        $this->assertInstanceOf(Person::class, $row['infos']['user']);
-        $this->assertEquals($row['infos']['score'], count($row['infos']['movies']));
-        $this->assertInstanceOf(Movie::class, $row['infos']['movies'][0]);
+        $this->assertInstanceOf(Person::class, $infos->get('user'));
+        $this->assertEquals($infos->get('score'), count($infos->get('movies')));
+        foreach ($infos->get('movies') as $movie) {
+            $this->assertInstanceOf(Movie::class, $movie);
+        }
     }
 }
