@@ -15,7 +15,7 @@ use GraphAware\Neo4j\OGM\Tests\Integration\Models\NodePropertyKeyMapping\Employe
 
 class NodePropertyKeyMappingTest extends IntegrationTestCase
 {
-    public function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->clearDb();
