@@ -17,7 +17,7 @@ use GraphAware\Neo4j\OGM\Tests\Integration\Models\RelationshipPropertyKeyMapping
 
 class RelationshipPropertyKeyMappingTest extends IntegrationTestCase
 {
-    public function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->clearDb();
