@@ -18,7 +18,7 @@ use GraphAware\Neo4j\OGM\Tests\Integration\Models\Base\User;
  */
 class RepositoryFinderFilterTest extends IntegrationTestCase
 {
-    public function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->clearDb();
